+++ apollo-portal/src/main/java/com/ctrip/framework/apollo/portal/controller/ServerConfigController.java
@@ -17,13 +17,10 @@
 package com.ctrip.framework.apollo.portal.controller;
 
 
-import com.ctrip.framework.apollo.common.utils.BeanUtils;
 import com.ctrip.framework.apollo.portal.entity.po.ServerConfig;
-import com.ctrip.framework.apollo.portal.repository.ServerConfigRepository;
+import com.ctrip.framework.apollo.portal.environment.Env;
 import com.ctrip.framework.apollo.portal.service.ServerConfigService;
-import com.ctrip.framework.apollo.portal.spi.UserInfoHolder;
 import java.util.List;
-import java.util.Objects;
 import javax.validation.Valid;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -37,46 +34,34 @@ import org.springframework.web.bind.annotation.RestController;
  */
 @RestController
 public class ServerConfigController {
-
-  private final ServerConfigRepository serverConfigRepository;
-  private final UserInfoHolder userInfoHolder;
   private final ServerConfigService serverConfigService;
 
-  public ServerConfigController(final ServerConfigRepository serverConfigRepository, final UserInfoHolder userInfoHolder, final ServerConfigService serverConfigService) {
-    this.serverConfigRepository = serverConfigRepository;
-    this.userInfoHolder = userInfoHolder;
+  public ServerConfigController(final ServerConfigService serverConfigService) {
     this.serverConfigService = serverConfigService;
   }
 
   @PreAuthorize(value = "@permissionValidator.isSuperAdmin()")
-  @PostMapping("/server/config")
-  public ServerConfig createOrUpdate(@Valid @RequestBody ServerConfig serverConfig) {
-    String modifiedBy = userInfoHolder.getUser().getUserId();
-
-    ServerConfig storedConfig = serverConfigRepository.findByKey(serverConfig.getKey());
-
-    if (Objects.isNull(storedConfig)) {//create
-      serverConfig.setDataChangeCreatedBy(modifiedBy);
-      serverConfig.setDataChangeLastModifiedBy(modifiedBy);
-      serverConfig.setId(0L);//为空，设置ID 为0，jpa执行新增操作
-      return serverConfigRepository.save(serverConfig);
-    }
-    //update
-    BeanUtils.copyEntityProperties(serverConfig, storedConfig);
-    storedConfig.setDataChangeLastModifiedBy(modifiedBy);
-    return serverConfigRepository.save(storedConfig);
+  @PostMapping("/server/portal-db/config")
+  public ServerConfig createOrUpdatePortalDBConfig(@Valid @RequestBody ServerConfig serverConfig) {
+    return serverConfigService.createOrUpdatePortalDBConfig(serverConfig);
   }
 
   @PreAuthorize(value = "@permissionValidator.isSuperAdmin()")
-  @GetMapping("/server/config/find-all-config")
-  public List<ServerConfig> findAllServerConfig() {
-    return serverConfigService.findAll();
+  @PostMapping("/server/envs/{env}/config-db/config")
+  public ServerConfig createOrUpdateConfigDBConfig(@Valid @RequestBody ServerConfig serverConfig, @PathVariable String env) {
+    return serverConfigService.createOrUpdateConfigDBConfig(Env.transformEnv(env), serverConfig);
   }
 
   @PreAuthorize(value = "@permissionValidator.isSuperAdmin()")
-  @GetMapping("/server/config/{key:.+}")
-  public ServerConfig loadServerConfig(@PathVariable String key) {
-    return serverConfigRepository.findByKey(key);
+  @GetMapping("/server/portal-db/config/find-all-config")
+  public List<ServerConfig> findAllPortalDBServerConfig() {
+    return serverConfigService.findAllPortalDBConfig();
+  }
+
+  @PreAuthorize(value = "@permissionValidator.isSuperAdmin()")
+  @GetMapping("/server/envs/{env}/config-db/config/find-all-config")
+  public List<ServerConfig> findAllConfigDBServerConfig(@PathVariable String env) {
+    return serverConfigService.findAllConfigDBConfig(Env.transformEnv(env));
   }
 
 }