+++ apollo-portal/src/main/resources/static/scripts/controller/ServerConfigController.js
@@ -14,51 +14,172 @@
  * limitations under the License.
  *
  */
-server_config_module.controller('ServerConfigController',
-    ['$scope', '$window', '$translate', 'toastr', 'ServerConfigService', 'AppUtil', 'PermissionService',
-        function ($scope, $window, $translate, toastr, ServerConfigService, AppUtil, PermissionService) {
+server_config_manage_module.controller('ServerConfigController',
+    ['$scope', '$window', '$translate', 'toastr', 'AppUtil', 'ServerConfigService', 'PermissionService',
+        'EnvService', ServerConfigController]);
 
-            $scope.serverConfig = {};
-            $scope.saveBtnDisabled = true;
+function ServerConfigController($scope, $window, $translate, toastr, AppUtil, ServerConfigService, PermissionService, EnvService) {
 
-            initPermission();
+    $scope.serverConfig = {};
+    $scope.portalDBConfigs = [];
+    $scope.portalDBFilterConfigs = [];
+    $scope.configDBConfigs = [];
+    $scope.configDBFilterConfigs = [];
+    $scope.envs = [];
+    $scope.selectedEnv = '';
+    $scope.displayModule = 'home';
+    $scope.portalDBConfigSearchKey = '';
+    $scope.configDBConfigSearchKey = '';
+    $scope.configEdit = configEdit;
+    $scope.createPortalDBConfig = createPortalDBConfig;
+    $scope.createConfigDBConfig = createConfigDBConfig;
+    $scope.gobackPortalDBTabs = gobackPortalDBTabs;
+    $scope.gobackConfigDBTabs = gobackConfigDBTabs;
+    $scope.portalDBConfigFilter = portalDBConfigFilter;
+    $scope.configDBConfigFilter = configDBConfigFilter;
+    $scope.resetPortalDBConfigSearchKey = resetPortalDBConfigSearchKey;
+    $scope.resetConfigDBConfigSearchKey = resetConfigDBConfigSearchKey;
+    $scope.switchConfigDBEnvs = switchConfigDBEnvs;
 
-            function initPermission() {
-                PermissionService.has_root_permission()
-                .then(function (result) {
-                    $scope.isRootUser = result.hasPermission;
-                })
-            }
+    $scope.allowSwitchingTabs = true;
 
-            $scope.create = function () {
-                ServerConfigService.create($scope.serverConfig).then(function (result) {
-                    toastr.success($translate.instant('ServiceConfig.Saved'));
-                    $scope.saveBtnDisabled = true;
-                    $scope.serverConfig = result;
-                }, function (result) {
-                    toastr.error(AppUtil.errorMsg(result), $translate.instant('ServiceConfig.SaveFailed'));
-                });
+    init();
+    function init() {
+        initPermission();
+        getPortalDBConfig();
+        initEnv();
+    }
+    function initEnv() {
+        EnvService.find_all_envs().then(function (result) {
+            $scope.envs = result;
+            $scope.selectedEnv = result[0];
+            getConfigDBConfig();
+        });
+    }
+    function initPermission() {
+        PermissionService.has_root_permission()
+        .then(function (result) {
+            $scope.isRootUser = result.hasPermission;
+        });
+    }
+
+    function getPortalDBConfig() {
+        ServerConfigService.findPortalDBConfig()
+        .then(function (result) {
+            $scope.portalDBConfigs = [];
+            $scope.portalDBFilterConfigs = [];
+            result.forEach(function (user) {
+                $scope.portalDBConfigs.push(user);
+                $scope.portalDBFilterConfigs.push(user);
+            });
+        },function (result) {
+            $scope.portalDBConfigs = [];
+            $scope.portalDBFilterConfigs = [];
+            toastr.error(AppUtil.errorMsg(result), $translate.instant('Config.SystemError'));
+        });
+    }
+
+    function getConfigDBConfig() {
+        ServerConfigService.findConfigDBConfig($scope.selectedEnv)
+        .then(function (result) {
+            $scope.configDBConfigs = [];
+            $scope.configDBFilterConfigs = [];
+            result.forEach(function (user) {
+                $scope.configDBConfigs.push(user);
+                $scope.configDBFilterConfigs.push(user);
+            });
+        },function (result) {
+            $scope.configDBConfigs = [];
+            $scope.configDBFilterConfigs = [];
+            toastr.error(AppUtil.errorMsg(result), $translate.instant('Config.SystemError'));
+        });
+    }
+
+    function configEdit (displayModule,config) {
+        $scope.displayModule = displayModule;
+        $scope.allowSwitchingTabs = false;
+
+        $scope.serverConfig = {};
+        if (config != null) {
+            $scope.serverConfig = {
+                key: config.key,
+                value: config.value,
+                comment: config.comment
             };
+        }
+    }
 
-            $scope.getServerConfigInfo = function () {
-                if (!$scope.serverConfig.key) {
-                    toastr.warning($translate.instant('ServiceConfig.PleaseEnterKey'));
-                    return;
-                }
+    function switchConfigDBEnvs(env) {
+        $scope.selectedEnv = env;
+        getConfigDBConfig();
+    }
 
-                ServerConfigService.getServerConfigInfo($scope.serverConfig.key).then(function (result) {
-                    $scope.saveBtnDisabled = false;
+    function createPortalDBConfig() {
+        ServerConfigService.createPortalDBConfig($scope.serverConfig).then(function (result) {
+            toastr.success($translate.instant('ServiceConfig.Saved'));
+            getPortalDBConfig();
+            $scope.displayModule = 'home';
+            $scope.allowSwitchingTabs = true;
+        }, function (result) {
+            toastr.error(AppUtil.errorMsg(result), $translate.instant('ServiceConfig.SaveFailed'));
+        });
+    }
 
-                    if (!result.key) {
-                        toastr.info($translate.instant('ServiceConfig.KeyNotExistsAndCreateTip', { key: $scope.serverConfig.key }));
-                        return;
-                    }
+    function createConfigDBConfig() {
+        ServerConfigService.createConfigDBConfig($scope.selectedEnv, $scope.serverConfig).then(function (result) {
+            toastr.success($translate.instant('ServiceConfig.Saved'));
+            getConfigDBConfig();
+            $scope.displayModule = 'home';
+            $scope.allowSwitchingTabs = true;
+        }, function (result) {
+            toastr.error(AppUtil.errorMsg(result), $translate.instant('ServiceConfig.SaveFailed'));
+        });
+    }
 
-                    toastr.info($translate.instant('ServiceConfig.KeyExistsAndSaveTip', { key: $scope.serverConfig.key }));
-                    $scope.serverConfig = result;
-                }, function (result) {
-                    AppUtil.showErrorMsg(result);
-                })
+
+    function gobackPortalDBTabs(){
+        $scope.displayModule = 'home';
+        $scope.allowSwitchingTabs = true;
+        getPortalDBConfig();
+    }
+
+    function gobackConfigDBTabs(){
+        $scope.displayModule = 'home';
+        $scope.allowSwitchingTabs = true;
+        getConfigDBConfig();
+    }
+
+    function portalDBConfigFilter() {
+        $scope.portalDBConfigSearchKey = $scope.portalDBConfigSearchKey.toLowerCase();
+        let filterConfig = [];
+        $scope.portalDBConfigs.forEach(function (item) {
+            let keyName = item.key;
+            if (keyName && keyName.toLowerCase().indexOf( $scope.portalDBConfigSearchKey) >= 0) {
+                filterConfig.push(item);
             }
+        });
+        $scope.portalDBFilterConfigs = filterConfig;
+    }
 
-        }]);
+    function resetPortalDBConfigSearchKey() {
+        $scope.portalDBConfigSearchKey = '';
+        portalDBConfigFilter();
+    }
+
+    function configDBConfigFilter() {
+        $scope.configDBConfigSearchKey = $scope.configDBConfigSearchKey.toLowerCase();
+        let filterConfig = [];
+        $scope.configDBConfigs.forEach(function (item) {
+            let keyName = item.key;
+            if (keyName && keyName.toLowerCase().indexOf( $scope.configDBConfigSearchKey) >= 0) {
+                filterConfig.push(item);
+            }
+        });
+        $scope.configDBFilterConfigs = filterConfig;
+    }
+
+    function resetConfigDBConfigSearchKey() {
+        $scope.configDBConfigSearchKey = '';
+        configDBConfigFilter();
+    }
+}