+++ apollo-portal/src/test/java/com/ctrip/framework/apollo/portal/controller/ServerConfigControllerTest.java
@@ -17,7 +17,9 @@
 package com.ctrip.framework.apollo.portal.controller;
 
 import com.ctrip.framework.apollo.portal.AbstractIntegrationTest;
+import com.ctrip.framework.apollo.portal.api.AdminServiceAPI;
 import com.ctrip.framework.apollo.portal.entity.po.ServerConfig;
+import com.ctrip.framework.apollo.portal.environment.Env;
 import com.ctrip.framework.apollo.portal.service.ServerConfigService;
 import org.junit.Assert;
 import org.junit.Test;
@@ -33,7 +35,8 @@ import java.util.*;
 import static org.hamcrest.core.StringContains.containsString;
 import static org.junit.Assert.assertEquals;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.mockito.Mockito.*;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.when;
 
 /**
  * Created by kezhenxu at 2019/1/14 13:24.
@@ -44,30 +47,31 @@ import static org.mockito.Mockito.*;
 public class ServerConfigControllerTest extends AbstractIntegrationTest {
   @Mock
   private ServerConfigService serverConfigService;
+
   @InjectMocks
   private ServerConfigController serverConfigController;
 
   @Test
   @Sql(scripts = "/sql/cleanup.sql", executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD)
-  public void shouldSuccessWhenParameterValid() {
+  public void shouldSuccessWhenParameterValidForPortalDBConfig() {
     ServerConfig serverConfig = new ServerConfig();
     serverConfig.setKey("validKey");
     serverConfig.setValue("validValue");
     ResponseEntity<ServerConfig> responseEntity = restTemplate.postForEntity(
-        url("/server/config"), serverConfig, ServerConfig.class
+        url("/server/portal-db/config"), serverConfig, ServerConfig.class
     );
     assertEquals(responseEntity.getBody().getKey(), serverConfig.getKey());
     assertEquals(responseEntity.getBody().getValue(), serverConfig.getValue());
   }
 
   @Test
-  public void shouldFailWhenParameterInvalid() {
+  public void shouldFailWhenParameterInvalidForPortalDBConfig() {
     ServerConfig serverConfig = new ServerConfig();
     serverConfig.setKey("  ");
     serverConfig.setValue("valid");
     try {
       restTemplate.postForEntity(
-          url("/server/config"), serverConfig, ServerConfig.class
+          url("/server/portal-db/config"), serverConfig, ServerConfig.class
       );
       Assert.fail("Should throw");
     } catch (final HttpClientErrorException e) {
@@ -80,7 +84,7 @@ public class ServerConfigControllerTest extends AbstractIntegrationTest {
     serverConfig.setValue("   ");
     try {
       restTemplate.postForEntity(
-          url("/server/config"), serverConfig, ServerConfig.class
+          url("/server/portal-db/config"), serverConfig, ServerConfig.class
       );
       Assert.fail("Should throw");
     } catch (final HttpClientErrorException e) {
@@ -93,9 +97,16 @@ public class ServerConfigControllerTest extends AbstractIntegrationTest {
 
   @Test
   public void testFindEmpty() {
-    when(serverConfigService.findAll()).thenReturn(new ArrayList<>());
-    List<ServerConfig> serverConfigList = serverConfigController.findAllServerConfig();
+    when(serverConfigService.findAllPortalDBConfig()).thenReturn(new ArrayList<>());
+    when(serverConfigService.findAllConfigDBConfig(Env.DEV)).thenReturn(new ArrayList<>());
+
+    List<ServerConfig> serverConfigList = serverConfigController.findAllPortalDBServerConfig();
     Assert.assertNotNull(serverConfigList);
     Assert.assertEquals(0, serverConfigList.size());
+
+    serverConfigList = serverConfigController.findAllConfigDBServerConfig(Env.DEV.getName());
+    Assert.assertNotNull(serverConfigList);
+    Assert.assertEquals(0, serverConfigList.size());
+
   }
 }